refactor(web): extract judge GroupRow cluster into pages/judge/GroupRow.tsx (PRD #1049)
